if/else to run the enter score page.
Debugged the incorrect/correct answer that was showing by moving down the "currentQuestion++;" part of my code to be below the if else statement. Made it so an if/else statments runs when the questions length in reached and it runs the enterHighScores page. Still need to work on the timer.