Created the highscores page
Created styling on the highscores page in css. The function showHighScores shows the initials on the leaderboard screen. Created the goBack button to refresh the page and takes you back to the start. Created the clearHighscores button to clear the locally stored data. Added the link viewHighScores to navigate to the leaderboard.