Hide score div if no scores present.
I added in the "if" statement on the highscores page so it doesn't show the div of the 1. TEST - 20 if you press the clear score button, by storing those values in a variable and to only run that if their values return true. Changed the text styling to text align left for the buttons.