Timer and final checks :)
Timer - created a setInterval timer to count down from 75 seconds inside the function called startQuiz. Added in my recordHighScore function a count = count - 10 in the incorrect if/else statement so the count decreases by 10 for each answer is wrong. Fixed up some spelling errors in the questions array. Added screenshots and final checks for the criteria :)